media-plugins/vapoursynth-curve: new plugin
This filter is similar to Adobe Photoshop and GIMP curves tools. Each plane has its values defined by N key points tied from each other using a smooth curve. The x-axis represents the pixel values from the input frame, and the y-axis the new pixel values to be set for the output frame.

By default, a curve is defined by the two points (0;0) and (1;1). This creates a straight line where each original pixel value is adjusted to its own value, which means no change to the image.

The filter allows you to redefine these two points and add some more. A new curve (using a natural cubic spline interpolation) will be defined to pass smoothly through all these new coordinates. The new defined points need to be strictly increasing over the x-axis, and their x and y values must be in the [0;1] interval. If the computed curves happened to go outside the vector spaces, the values will be clipped accordingly.