media-plugins/vapoursynth-median: new plugin
Median generates a pixel-by-pixel median of several clips. This is particularly useful for filtering out noise and glitches from multiple VHS/SVHS/8mm/Hi8 tape captures, but can be used for other purposes also.

TemporalMedian returns the temporal median of a single clip.

MedianBlend can return a clip derived from the minimum or maximum pixel values, or it can discard some low (default: 1) and high (default: 1) values and blend the others together. With the parameters set to not discard anything, the result is in fact the average of the clips.

This is a port of the Avisynth plugin Median, version 0.6.