media-plugins/vs-ccd: new plugin
CCD is a simple chroma denoiser. It works by selectively averaging pixels in a 25x25 matrix below the Euclidean distance threshold in an RGB clip. After denoising, the clip should be converted back to YUV / YCoCg, and the luma channel should be copied from the input.

Currently, CCD only supports RGBS input and will not copy the original luma for you, so for convinience, a python wrapper is included which can handle this.

Vapoursynth port of the original filter by Sergey Stolyarevsky.