media-plugins/bestsource: new plugin
BestSource (abbreviated as BS) is a cross-platform wrapper library around FFmpeg that ensures sample/frame accurate access to audio and video by always linearly decoding the input files. With a few tricks this can guarantee accurate seeking and can be surprisingly... not slow.

It can be used as either a C++ library directly or through the VapourSynth plugin that's included.